Customauth/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-guardian==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
future==0.18.2
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2022.1
requests==2.28.0
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
